fix: zero, init, copy, minusをqd_util.cppに分割
